+++ src/Http/Requests/ResourceRequest.php
@@ -224,19 +224,25 @@ class ResourceRequest extends FormRequest
     }
 
     /**
-     * @inheritDoc
+     * Prepare the data for validation.
+     *
+     * Before validating the data, we need to:
+     *
+     * 1. Ensure the media type is supported;
+     * 2. Parse the document for compliance with the JSON:API spec.
+     *
+     * We only need to do this for the requests where we are expecting a JSON:API document,
+     * i.e. resource create/update requests and requests to modify a resource's relationship.
+     *
+     * @return void
      */
     protected function prepareForValidation()
     {
-        /** Content negotiation. */
-        if (!$this->isSupportedMediaType()) {
-            throw $this->unsupportedMediaType();
-        }
-
-        /** JSON API spec compliance. */
         if ($this->isCreating() || $this->isUpdating()) {
+            $this->assertSupportedMediaType();
             $this->validateResourceDocument();
         } else if ($this->isModifyingRelationship()) {
+            $this->assertSupportedMediaType();
             $this->validateRelationshipDocument();
         }
     }
@@ -417,6 +423,16 @@ class ResourceRequest extends FormRequest
         }
 
         return $data;
+    }
+
+    /**
+     * @throws HttpExceptionInterface
+     */
+    private function assertSupportedMediaType(): void
+    {
+        if (!$this->isSupportedMediaType()) {
+            throw $this->unsupportedMediaType();
+        }
     }
 
     /**